+++ data/CHANGELOG.md
@@ -1,5 +1,9 @@
 # RaisedCurious Changelog
 
+## 2026-05-04 (Monday)
+
+- Experiment "Soap Film Color Orchestra" (#auto-2026-05-04)
+
 ## 2026-04-27 (Monday)
 
 - Experiment "Magnetic Cereal: Iron in Your Breakfast" (#auto-2026-04-27)